extracted client logic out into separate packages for testing purposes
